Make Like a seperate componant
